third_party/blink/web_tests/fast/dom/HTMLImageElement/image-picture-react-to-media-changes-expected.txt

PASS -1 is not img.currentSrc.indexOf("image-set-4x")
PASS -1 is not img.currentSrc.indexOf("blue_rect")
PASS successfullyParsed is true

TEST COMPLETE
